[Refactor/#26] 피드백 반영
