+++ src/view/main/form/todoForm.css
@@ -36,6 +36,8 @@
   border-radius: 4px;
 
   font-size: 1rem;
+
+  cursor: cell;
 }
 
 .add-btn:hover,
