+++ src/view/main/todoCards/todoCards.css
@@ -11,8 +11,10 @@
 /* 삭제, 완료 / 취소, 저장버튼 감싸는 틀 */
 .delete-nd-done,
 .cancel-nd-save {
+  width: 100%;
   position: sticky;
   bottom: 0px;
+  right: 0px;
   display: flex;
   justify-content: space-between;
 }
@@ -25,6 +27,7 @@
   border-radius: 10px;
 
   font-size: 1rem;
+  cursor: cell;
 }
 
 .delete-btn:hover,
@@ -40,6 +43,7 @@
   border-radius: 10px;
 
   font-size: 1rem;
+  cursor: cell;
 }
 
 /* 완료버튼과 저장버튼 */
@@ -53,6 +57,8 @@
   font-size: 1rem;
   background-color: #333;
   color: white;
+
+  cursor: cell;
 }
 
 .done-btn:hover,
@@ -62,8 +68,11 @@
 }
 
 .edit-wrap {
+  min-width: 200px;
+  max-width: 100%;
   display: flex;
   flex-direction: column;
+
   gap: 12px;
 }
 
@@ -75,8 +84,6 @@
 }
 
 .edit-wrap > textarea {
-  min-width: 300px;
-  max-width: 100%;
   padding: 6px;
 
   font-size: 1rem;
@@ -90,6 +97,8 @@
 
   font-size: 1rem;
   background-color: #fff;
+
+  cursor: cell;
 }
 
 .edit-btn:hover {
